+++ package.json
@@ -26,6 +26,7 @@
     "react-hook-form": "^7.54.2",
     "react-icons": "^5.4.0",
     "react-redux": "^9.2.0",
+    "redux-persist": "^6.0.0",
     "styled-components": "^6.1.14"
   },
   "devDependencies": {
