+++ src/app/api/users/[loginuser]/route.js
@@ -1,0 +1,8 @@
+import { NextResponse } from 'next/server'
+ 
+export async function GET(request) {
+   const loginUserToken = request.headers.get("Authorization");
+console.log(loginUserToken);
+
+  return NextResponse.json({ Result: loginUserToken},{status:200})
+}
