--- src/app/api/users/user.js/route.js
@@ -1,5 +1,0 @@
-import { NextResponse } from 'next/server'
- 
-export async function GET() {
-  return NextResponse.json({ Result: "hello"},{status:200})
-}
