+++ src/components/Chat/Chat.js
@@ -5,20 +5,17 @@ import { BsSendFill } from "react-icons/bs";
 import { useDispatch, useSelector } from "react-redux";
 
 function Chat() {
-function handleclick(){
- 
-}
+
 
   return (
     <section className="bg-[#4a2184] text-white min-h-screen flex items-center justify-center">
     <div className="container p-4">
       <div className="grid grid-cols-12 rounded-lg bg-[#391965] gap-4">
-        {/* Left Sidebar */}
         <div className="col-span-4 p-4 rounded-lg flex flex-col space-y-4">
           <h2 className="text-lg font-semibold">Chat Users</h2>
           <ul className="space-y-2">
             <li className="p-3 rounded-lg bg-[#4a2184] cursor-pointer hover:bg-[#614aa8]">
-              User {count}
+              User 
             </li>
             <li className="p-3 rounded-lg bg-[#4a2184] cursor-pointer hover:bg-[#614aa8]">
               User 2
@@ -32,10 +29,8 @@ function handleclick(){
           </ul>
         </div>
   
-        {/* Main Chat Area */}
         <div className="col-span-8 p-4 bg-[#4a2184] rounded-lg flex flex-col">
           <div className="flex-1 overflow-y-auto p-4 bg-[#391965] rounded-lg">
-            {/* Chat messages */}
             <div className="space-y-3">
               <div className="bg-[#614aa8] p-3 rounded-lg self-start max-w-xs">
                 Hello! How are you?
@@ -46,7 +41,6 @@ function handleclick(){
             </div>
           </div>
   
-          {/* Chat Input */}
           <form className="mt-4 flex items-center space-x-2">
             <input
               type="text"
@@ -54,8 +48,6 @@ function handleclick(){
               placeholder="Type a message..."
             />
             <span
-            onClick={handleclick}
-              
               className="p-3 items-center gap-2 flex bg-[#614aa8] rounded-lg hover:bg-[#814ac8] transition"
             >
               Send
