+++ src/components/User/FromHeader.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+function FromHeader() {
+  return (
+    <div className="text-3xl max-w-fit text-center p-10">
+    <h3 className="font-medium pb-2"> Create new Account</h3>
+    <p className="text-lg text-slate-400">
+      Create to your account and find your mate with us!
+    </p>
+  </div>
+  )
+}
+
+export default FromHeader
